test exclusion of code blocks that do not specify javascript
Despite Rollup being a JavaScript-specific build tool, it's still useful to require the use of fences with language annotation a la ```js and ```javascript because that makes it possible to also include illustrative code in other languages, in particular Bash installation (npm install) or JSON data structures. Add unit tests to make sure that fences without language annotations will not be included in the output.